util/time: move SlowTimer to mod time
