+++ src/raftstore/store/store.rs
@@ -32,8 +32,8 @@ use kvproto::raft_serverpb::{RaftMessage, RaftSnapshotData, RaftTruncatedState, 
                              PeerState};
 use kvproto::eraftpb::{ConfChangeType, MessageType};
 use kvproto::pdpb::StoreStats;
-use util::{SlowTimer, escape};
-use util::time::duration_to_sec;
+use util::escape;
+use util::time::{SlowTimer, duration_to_sec};
 use pd::PdClient;
 use kvproto::raft_cmdpb::{AdminCmdType, AdminRequest, StatusCmdType, StatusResponse,
                           RaftCmdRequest, RaftCmdResponse};
