+++ src/util/mod.rs
@@ -16,7 +16,7 @@ use std::ops::DerefMut;
 use std::io;
 use std::{slice, thread};
 use std::net::{ToSocketAddrs, TcpStream, SocketAddr};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 use std::collections::hash_map::Entry;
 use std::sync::{RwLock, RwLockReadGuard, RwLockWriteGuard};
 use std::collections::vec_deque::{Iter, VecDeque};
@@ -253,48 +253,6 @@ pub fn as_slice<T>(t: &T) -> &[T] {
     unsafe {
         let ptr = t as *const T;
         slice::from_raw_parts(ptr, 1)
-    }
-}
-
-pub struct SlowTimer {
-    slow_time: Duration,
-    t: Instant,
-}
-
-impl SlowTimer {
-    pub fn new() -> SlowTimer {
-        SlowTimer::default()
-    }
-
-    pub fn from(slow_time: Duration) -> SlowTimer {
-        SlowTimer {
-            slow_time: slow_time,
-            t: Instant::now(),
-        }
-    }
-
-    pub fn from_secs(secs: u64) -> SlowTimer {
-        SlowTimer::from(Duration::from_secs(secs))
-    }
-
-    pub fn from_millis(millis: u64) -> SlowTimer {
-        SlowTimer::from(Duration::from_millis(millis))
-    }
-
-    pub fn elapsed(&self) -> Duration {
-        self.t.elapsed()
-    }
-
-    pub fn is_slow(&self) -> bool {
-        self.elapsed() >= self.slow_time
-    }
-}
-
-const DEFAULT_SLOW_SECS: u64 = 1;
-
-impl Default for SlowTimer {
-    fn default() -> SlowTimer {
-        SlowTimer::from_secs(DEFAULT_SLOW_SECS)
     }
 }
